스프레드 시트 연결 - 유닛

+++ Assets/01_Scripts/01_Unit/RewardManager.cs
@@ -6,7 +6,7 @@ public class RewardManager : MonoBehaviour
 {
     public static RewardManager Instance { get; set; }
     public int Gold { get; set; }
-    public int Exp { get; set; }
+    public float Exp { get; set; }
 
     private void Awake()
     {
